Add F1 implementation

+++ pkg/utils/utils.go
@@ -13,17 +13,17 @@ func Set(x uint) []uint {
 
 // Concat performs zero-padded concatenation of x and y
 // y belongs to [2^k]
-func Concat(x, y *big.Int, k uint) *big.Int {
-	return x.Lsh(x, k).Add(x, y)
+func Concat(x, y *big.Int, k uint64) *big.Int {
+	return x.Lsh(x, uint(k)).Add(x, y)
 }
 
 // Trunc returns the b most significant of x. If a is non-zero then the ath to (b-1)th
 // bits of x are returned. x belongs to [2^k]
-func Trunc(x *big.Int, a, b, k uint) *big.Int {
-	x.Rsh(x, k-b)
+func Trunc(x *big.Int, a, b, k uint64) *big.Int {
+	x.Rsh(x, uint(k-b))
 	if a > 0 {
 		least := big.NewInt(1)
-		least.Lsh(least, a)
+		least.Lsh(least, uint(a))
 		x.Mod(x, least)
 	}
 	return x
